separate styled HOCs from SelectField
